[UPD]'customer_area': actualizado módulo a versión 11
