+++ project-addons/customer_area/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 ##############################################################################
 #
 #    Copyright (C) 2014 Pexego Sistemas Informáticos All Rights Reserved
@@ -19,7 +18,5 @@
 #
 ##############################################################################
 
-from . import res_partner
-from . import sale
+from . import models
 from . import report
-from . import res_users
