+++ project-addons/customer_area/__manifest__.py
@@ -19,12 +19,21 @@
 #
 ##############################################################################
 
-from openerp import models, fields
-
-
-class sale_order(models.Model):
-
-    _inherit = 'sale.order'
-
-    area_id = fields.Many2one('res.partner.area', string='Area', store=True,
-                              related='partner_id.area_id')
+{
+    'name': "Customer area",
+    'version': '1.0',
+    'category': 'sale',
+    'description': """Add area to partner""",
+    'author': 'Pexego Sistemas Informáticos',
+    'website': 'www.pexego.es',
+    "depends" : ['base',
+                 'sale',
+                 ],
+    "data" : ['views/res_partner_view.xml',
+              'views/partner_area_view.xml',
+              'report/sale_report_view.xml',
+              'views/sale_view.xml',
+              'security/ir.model.access.csv',
+              'views/res_users_view.xml'],
+    "installable": True
+}
