+++ project-addons/customer_area/models/__init__.py
@@ -1,0 +1,3 @@
+from . import res_partner
+from . import sale
+from . import res_users
