+++ project-addons/customer_area/models/res_users.py
@@ -1,8 +1,7 @@
-# -*- coding: utf-8 -*-
 ##############################################################################
 #
-#    Copyright (C) 2014 Pexego Sistemas Informáticos All Rights Reserved
-#    $Jesús Ventosinos Mayor <jesus@pexego.es>$
+#    Copyright (C) 2016 Comunitea Servicios Tecnológicos All Rights Reserved
+#    $Omar Castiñeira Saavedra <omar@comunitea.com>$
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as published
@@ -19,21 +18,12 @@
 #
 ##############################################################################
 
-{
-    'name': "Customer area",
-    'version': '1.0',
-    'category': 'sale',
-    'description': """Add area to partner""",
-    'author': 'Pexego Sistemas Informáticos',
-    'website': 'www.pexego.es',
-    "depends" : ['base',
-                 'sale',
-                 ],
-    "data" : ['res_partner_view.xml',
-              'partner_area_view.xml',
-              'report/sale_report_view.xml',
-              'sale_view.xml',
-              'security/ir.model.access.csv',
-              'res_users_view.xml'],
-    "installable": True
-}
+from odoo import models, fields
+
+
+class ResUsers(models.Model):
+
+    _inherit = 'res.users'
+
+    area_ids = fields.Many2many('res.partner.area', 'res_users_area_rel',
+                                'user_ud', 'area_id', 'Areas')
