+++ project-addons/customer_area/views/partner_area_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
 <data>
 
     <record id="res_partner_area_tree" model="ir.ui.view">
@@ -52,7 +52,7 @@
         <field name="search_view_id" ref="res_partner_area_filter"/>
     </record>
 
-    <menuitem name="Areas" parent="base.menu_base_config" action="action_res_partner_area"
+    <menuitem name="Areas" parent="sale.sales_team_config" action="action_res_partner_area"
               id="res_partner_area_menu" sequence="9"/>
 
      <record id="res_partner_area_region_tree" model="ir.ui.view">
@@ -105,8 +105,8 @@
         <field name="search_view_id" ref="res_partner_area_region_filter"/>
     </record>
 
-    <menuitem name="Commercial Regions" parent="base.menu_base_config" action="action_res_partner_area_region"
+    <menuitem name="Commercial Regions" parent="sale.sales_team_config" action="action_res_partner_area_region"
               id="res_partner_area_region_menu" sequence="9"/>
 
 </data>
-</openerp>
+</odoo>
