+++ project-addons/customer_area/views/res_partner_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
     <data>
 
         <record id="res_partner_view_area_search" model="ir.ui.view">
@@ -11,9 +11,9 @@
                     <field name="region_ids" string="Commercial Region"
                            filter_domain="['|', ('region_ids.code','ilike',self), ('region_ids','ilike',self)]"/>
                 </field>
-                <filter string="Country"  position="after">
-                   <filter string="Area" context="{'group_by': 'area_id'}"/>
-                </filter>
+                <xpath expr="//group[last()]" position="after">
+                    <filter string="Area" context="{'group_by': 'area_id'}"/>
+                </xpath>
             </field>
         </record>
 
@@ -23,8 +23,11 @@
             <field name="inherit_id" ref="base.view_partner_form" />
             <field name="arch" type="xml">
                 <field name="website" position="before">
-                    <field name="area_id" on_change="change_sales_team(area_id)"
+                    <!--ACTIVAR AL MIGRAR CUSTOM_PARTNER field name="area_id" on_change="change_sales_team(area_id)"
                            attrs="{'required': [('is_company', '=', True),('customer', '=', True),('dropship', '=', False)], 'invisible':[('customer', '=', False)]}"
+                           options="{'no_quick_create':True,'no_create_edit':True}"/-->
+                    <field name="area_id" on_change="change_sales_team(area_id)"
+                           attrs="{'required': [('is_company', '=', True),('customer', '=', True)], 'invisible':[('customer', '=', False)]}"
                            options="{'no_quick_create':True,'no_create_edit':True}"/>
                 </field>
                 <xpath expr="//field[@name='child_ids']/form//field[@name='mobile']" position="after">
@@ -36,7 +39,7 @@
                 <xpath expr="//field[@name='child_ids']/form//field[@name='mobile']" position="attributes">
                     <attribute name="string">Phone/Mobile 2</attribute>
                 </xpath>
-                <xpath expr="//field[@name='child_ids']/form//field[@name='use_parent_address']" position="before">
+                <xpath expr="//field[@name='child_ids']/form//field[@name='country_id']" position="after">
                     <div attrs="{'invisible': [('type','!=', 'delivery')]}" name="div_default">
                         <field name="default_shipping_address"/><label for="default_shipping_address"/>
                     </div>
@@ -69,4 +72,4 @@
         </record>
 
     </data>
-</openerp>
+</odoo>
