+++ project-addons/customer_area/views/res_users_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
     <data>
 
         <record id="view_users_form_add_area_ids" model="ir.ui.view">
@@ -7,13 +7,13 @@
             <field name="model">res.users</field>
             <field name="inherit_id" ref="base.view_users_form"/>
             <field name="arch" type="xml">
-                <group string="Allowed Companies" position="after">
-                    <group string="Work areas">
+                <field name="company_ids" position="after">
+                    <group string="Work areas" name="work_areas">
                         <field name="area_ids"/>
                     </group>
-                </group>
+                </field>
             </field>
         </record>
 
     </data>
-</openerp>
+</odoo>
